refactor: TagItem, AssigneeItem dto 중복 개선
