+++ backend/flowmeet-api/src/main/java/kr/flowmeet/api/node/dto/response/AssigneeItem.java
@@ -1,0 +1,15 @@
+package kr.flowmeet.api.node.dto.response;
+
+import kr.flowmeet.domain.user.entity.User;
+
+public record AssigneeItem(
+        Long userId,
+        String nickname,
+        String email,
+        String profileImageUrl
+) {
+
+    public static AssigneeItem from(final User user) {
+        return new AssigneeItem(user.getId(), user.getNickname(), user.getEmail(), user.getProfileImageUrl());
+    }
+}
