+++ backend/flowmeet-api/src/main/java/kr/flowmeet/api/node/dto/response/GetAllTagsResponse.java
@@ -8,11 +8,4 @@ public record GetAllTagsResponse(List<TagItem> tags) {
     public static GetAllTagsResponse from(final List<Tag> tags) {
         return new GetAllTagsResponse(tags.stream().map(TagItem::from).toList());
     }
-
-    public record TagItem(Long tagId, String name, String color) {
-
-        public static TagItem from(final Tag tag) {
-            return new TagItem(tag.getId(), tag.getName(), tag.getColor());
-        }
-    }
 }
