+++ backend/flowmeet-api/src/main/java/kr/flowmeet/api/node/dto/response/GetKanbanResponse.java
@@ -63,25 +63,4 @@ public record GetKanbanResponse(
             );
         }
     }
-
-    public record TagItem(
-            Long tagId,
-            String name,
-            String color
-    ) {
-
-        public static TagItem from(final Tag tag) {
-            return new TagItem(tag.getId(), tag.getName(), tag.getColor());
-        }
-    }
-
-    public record AssigneeItem(
-            Long userId,
-            String nickname
-    ) {
-
-        public static AssigneeItem from(final User user) {
-            return new AssigneeItem(user.getId(), user.getNickname());
-        }
-    }
 }
