+++ backend/flowmeet-api/src/main/java/kr/flowmeet/api/node/dto/response/GetNodeListResponse.java
@@ -61,26 +61,4 @@ public record GetNodeListResponse(
             );
         }
     }
-
-    public record TagItem(
-            Long tagId,
-            String name,
-            String color
-    ) {
-
-        public static TagItem from(final Tag tag) {
-            return new TagItem(tag.getId(), tag.getName(), tag.getColor());
-        }
-    }
-
-    public record AssigneeItem(
-            Long userId,
-            String nickname,
-            String profileImageUrl
-    ) {
-
-        public static AssigneeItem from(final User user) {
-            return new AssigneeItem(user.getId(), user.getNickname(), user.getProfileImageUrl());
-        }
-    }
 }
