+++ backend/flowmeet-api/src/main/java/kr/flowmeet/api/node/dto/response/SearchNodeResponse.java
@@ -40,15 +40,4 @@ public record SearchNodeResponse(
             );
         }
     }
-
-    public record TagItem(
-            Long tagId,
-            String name,
-            String color
-    ) {
-
-        public static TagItem from(final Tag tag) {
-            return new TagItem(tag.getId(), tag.getName(), tag.getColor());
-        }
-    }
 }
