+++ backend/flowmeet-api/src/main/java/kr/flowmeet/api/node/dto/response/TagItem.java
@@ -1,0 +1,10 @@
+package kr.flowmeet.api.node.dto.response;
+
+import kr.flowmeet.domain.node.entity.Tag;
+
+public record TagItem(Long tagId, String name, String color) {
+
+    public static TagItem from(final Tag tag) {
+        return new TagItem(tag.getId(), tag.getName(), tag.getColor());
+    }
+}
